created a shared types file
