+++ app/react/V2/Routes/Settings/IX/IXDashboard.tsx
@@ -9,8 +9,8 @@ import { SettingsContent } from 'V2/Components/Layouts/SettingsContent';
 import { ClientTemplateSchema } from 'app/istore';
 import { Button, ConfirmationModal, Table } from 'V2/Components/UI';
 import { Translate, t } from 'app/I18N';
-import { notificationAtom } from 'app/V2/atoms';
-import { IXExtractorInfo } from './types';
+import { notificationAtom } from 'V2/atoms';
+import { IXExtractorInfo } from 'V2/shared/types';
 import { ExtractorModal } from './components/ExtractorModal';
 import { Extractor, tableColumns } from './components/TableElements';
 import { List } from './components/List';
