+++ app/react/V2/shared/types.ts
@@ -1,0 +1,8 @@
+interface IXExtractorInfo {
+  _id?: string;
+  name: string;
+  property: string;
+  templates: string[];
+}
+
+export type { IXExtractorInfo };
